+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [TD-5336] New `i18n` service for managing i18n locales and messages
+- [TD-4711] Startup task to delete deprecated messages
 
 ### Changed